backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic
openai
pdfminer.six
pytube
tesseract
pillow
python-dotenv
aiofiles
